requirements.txt
python-telegram-bot==13.7
pymongo==3.11.4
python-dotenv==0.19.0
apscheduler==3.6.3